+++ push-notifications/src/index.ts
@@ -1,0 +1,10 @@
+import { registerPlugin } from '@capacitor/core';
+
+import type { PushNotificationsPlugin } from './definitions';
+
+const PushNotifications = registerPlugin<PushNotificationsPlugin>(
+  'PushNotifications',
+  {},
+);
+
+export { PushNotifications };