+++ dialog/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/dialog@0.2.3...@capacitor/dialog@0.2.4) (2021-01-13)
+
+**Note:** Version bump only for package @capacitor/dialog
+
+
+
+
+
 ## [0.2.3](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/dialog@0.2.2...@capacitor/dialog@0.2.3) (2021-01-08)
 
 **Note:** Version bump only for package @capacitor/dialog