+++ keyboard/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/keyboard@0.3.2...@capacitor/keyboard@0.3.3) (2021-01-13)
+
+**Note:** Version bump only for package @capacitor/keyboard
+
+
+
+
+
 ## [0.3.2](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/keyboard@0.3.1...@capacitor/keyboard@0.3.2) (2021-01-08)
 
 **Note:** Version bump only for package @capacitor/keyboard